+++ lib/generators/rails_browscap/install/templates/browscap.rb
@@ -1,0 +1,4 @@
+require 'browscap'
+module Rails
+  @@browscap = ::Browscap.new 'config/browscap.ini'
+end