+++ test/test_helper.rb
@@ -1,3 +1,6 @@
 require 'rubygems'
 require 'active_support'
 require 'active_support/test_case'
+
+require 'rails-browscap'
+require 'generators/rails_browscap/install/install_generator'