+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-* IN PROGRESS (0.1.0)
+* 0.1.0
   * inifile 0.4.0 to fix encoding with ruby 1.9
   * better description on Known issue
   * fixed versioning so I have a proper revision number
+  * generators / tasks work
+  * Initialize and define current_browser in the plugin itself
 
  * 0.0.4
   * fixed rake task